Refactor NPC Hierarchy
Changelog:
- 1st Level Class: AbstractNPC
- 2nd Level Classes: AbstractNeutral/Passive/HostileNPC
- 3rd Level Classes: AbstractPirate/NavyNPC
- 4th Level Classes: NavyNPC, PirateNPC, NPC, etc.

Documented code.
