+++ src/main/java/net/minemod/onepiecemod/client/pirate/PirateNPCRenderer.java
@@ -5,7 +5,7 @@ import net.minecraft.Util;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
 import net.minemod.onepiecemod.OnePieceMod;
-import net.minemod.onepiecemod.entity.npcs.pirate.PirateNPC;
+import net.minemod.onepiecemod.entity.npcs.neutral.pirate.PirateNPC;
 
 import java.util.Map;
